handle required request

+++ app/Services/TodoServices.php
@@ -2,6 +2,7 @@
 
 namespace App\Services;
 
+use App\Exceptions\TodoNotFoundException;
 use App\Models\Todo;
 use App\Repositories\TodoRepository;
 
@@ -14,16 +15,28 @@ class TodoServices
 
     public static function updateTodo(int $id, string $name): Todo
     {
+        $todo = Todo::find($id);
+        if (!$todo) {
+            throw new TodoNotFoundException('TODO not found by ID ', $todo);
+        }
         return TodoRepository::updateTodo($id, $name);
     }
 
     public static function completeTodo(int $id): Todo
     {
+        $todo = Todo::find($id);
+        if (!$todo) {
+            throw new TodoNotFoundException('TODO not found by ID ', $todo);
+        }
         return TodoRepository::completeTodo($id);
     }
 
     public static function deleteTodo(int $id): bool
     {
+        $todo = Todo::find($id);
+        if (!$todo) {
+            throw new TodoNotFoundException('TODO not found by ID ', $todo);
+        }
         return TodoRepository::deleteTodo($id);
     }
 }
